Tests now run in about 25 minutes.
