+++ .github/workflows/run_tests.yml
@@ -45,4 +45,4 @@ jobs:
       - name: Test with pytest
         run: |
           python -m pip install .
-          python -m pytest
+          python -m tests
