+++ lineax/_misc.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import functools as ft
-from typing import Callable
+from collections.abc import Callable
 
 import equinox as eqx
 import jax
