+++ lineax/_solver/bicgstab.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any, Callable, Optional
+from collections.abc import Callable
+from typing import Any, Optional
 from typing_extensions import TypeAlias
 
 import jax
