+++ lineax/_solver/cg.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Any, Callable, ClassVar, Optional
+from collections.abc import Callable
+from typing import Any, ClassVar, Optional
 from typing_extensions import TypeAlias
 
 import equinox.internal as eqxi
