+++ tests/README.md
@@ -1,0 +1,3 @@
+Each file is run separately to avoid JAX out-of-memory'ing.
+
+As such, run tests using `python -m tests`, *not* by just running `pytest`.
