+++ tests/test_operator.py
